Adopt commit d3033357d from master: Merge pull request #1051 from jmendelsohn4/drqs-98224955/bcema_nb_20170322
commit d3033357d56bf30380fde9dc62b40146d3a34282
Merge: 975c7ee3f 17e7dfaa0
Author: Hyman Rosen <hrosen4@bloomberg.net>
Date:   Wed Mar 22 15:02:28 2017 -0400

    Merge pull request #1051 from jmendelsohn4/drqs-98224955/bcema_nb_20170322

    fixing nightly build issues

commit 17e7dfaa06087d5802b73bb65c1cfa1120c71530
Author: Jeffrey Mendelsohn <JMENDELSOHN4@bloomberg.net>
Date:   Wed Mar 22 14:58:57 2017 -0400

    fixing nightly build issues

+++ groups/bal/balxml/balxml_formatter.h
@@ -671,7 +671,8 @@ void Formatter::addAttribute(const bslstl::StringRef& name,
                               formattingMode,
                               &d_encoderOptions);
         d_outputStream << '"';
-        d_column += name.length() + 4;  // Minimum output if value is empty
+        d_column += static_cast<int>(name.length()) + 4;
+                                           // Minimum output if value is empty.
     }
 }
